+++ src/main/java/util/DataUtils.java
@@ -1,11 +1,14 @@
 package util;
 
+import constants.Cookie;
 import org.junit.jupiter.params.provider.Arguments;
 
 import java.util.stream.Stream;
 
 import static constants.ResultValues.*;
-import static constants.ResultValues.BLOUSE_RESULT;
+import static constants.TestDataGeneratorForAddingToCartWithProperties.generateIpa;
+import static constants.TestDataGeneratorForAddingToCartWithProperties.generateQTY;
+import static constants.TestDataGeneratorForRegistration.*;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 public class DataUtils {
@@ -38,5 +41,29 @@ public class DataUtils {
         );
     }
 
+    static Stream<Arguments> provideEmail() {
+        return Stream.generate(() ->
+                arguments(Cookie.cookie1, emailGenerator())).limit(10);
+    }
+
+    static Stream<Arguments> provideArgumentsForAddingToCartWithDifSizeAndColor() {
+        return Stream.generate(() ->
+                arguments(generateQTY(), generateIpa())).limit(10);
+    }
+
+    static Stream<Arguments> provideArgumentsForAccountCreation() {
+        return Stream.generate(() ->
+                arguments(
+                        emailGenerator(),
+                        passwordGenerator(),
+                        fNameGenerator(),
+                        lNameGenerator(),
+                        addressGenerator(),
+                        cityGenerator(),
+                        zipCodeGenerator(),
+                        phoneNumberGenerator(),
+                        Cookie.cookie1
+                )).limit(5);
+    }
 
 }
